Knock: (HACK) Stage messaging/version.js
This is an annoying hack to make using our repo more seamless.

The issue was that messaging/lib/index.js couldn't find version.js in
the same folder, eventually I found out that each package has in its
package.json the build script which uses genversion to create this file.

So, why does it work in the real deal but not for us? Two reasons:
1. lerna (the build management tool) runs the script before publishing
   it and publishes it with the version files (for all packages)
2. The version files are ignored by git, so maintainers don't commit
   them

So, to fix that we'll have to remove the line from .gitignore and create
the file by ourselves (and commit it) for any packages we maintain.

So to upgrade the fork:
0. git checkout main
1. git pull --rebase package main
2. cd packages/messaging && yarn build
3. git tag -f v`python -c "import json; f = open('lerna.json'); j = json.load(f); print(j['version'])"`
4. git commit --amend (should be into this commit..)
5. git push --tags -f
6. git push -f

You can run `./update_version.sh` to do it automatically